+++ dagster-cloud/dagster_cloud/pex/grpc/server/cli/__init__.py
@@ -1,10 +1,13 @@
-import logging
 import os
-from typing import Optional
+import subprocess
+from typing import Optional, cast
 
-from dagster._core.utils import coerce_valid_log_level
+from dagster._serdes import deserialize_as
+from dagster._utils.interrupts import capture_interrupts
+from dagster_cloud_cli.core.workspace import PexMetadata
 from typer import Option, Typer, echo
 
+from ..registry import PexS3Registry
 from ..server import run_multipex_server
 
 app = Typer(hidden=True)
@@ -28,3 +31,29 @@ def grpc(
         max_workers=max_workers,
         local_pex_files_dir=local_pex_files_dir,
     )
+
+
+@app.command(short_help="Execute a run with a PEX file")
+def execute_run(
+    input_json: str,
+    pex_metadata_json: str,
+    local_pex_files_dir: Optional[str] = Option(
+        default="/tmp/pex-files", envvar="LOCAL_PEX_FILES_DIR"
+    ),
+):
+    with capture_interrupts():
+        pex_metadata = deserialize_as(pex_metadata_json, PexMetadata)
+        executable = PexS3Registry(local_pex_files_dir).get_pex_executable(pex_metadata)
+
+        subprocess.run(
+            [
+                executable.source_path,
+                "-m",
+                "dagster",
+                "api",
+                "execute_run",
+                input_json,
+            ],
+            check=True,
+            env={**os.environ.copy(), **executable.environ},
+        )