+++ dagster-cloud/dagster_cloud/storage/tags.py
@@ -1,3 +1,8 @@
-SYSTEM_TAG_PREFIX = "dagster-cloud/"
+from dagster._core.storage.tags import HIDDEN_TAG_PREFIX, SYSTEM_TAG_PREFIX
 
-ALERT_EMAILS_TAG = f"{SYSTEM_TAG_PREFIX}alert_emails"
+CLOUD_SYSTEM_TAG_PREFIX = "dagster-cloud/"
+
+ALERT_EMAILS_TAG = f"{CLOUD_SYSTEM_TAG_PREFIX}alert_emails"
+
+PEX_METADATA_TAG = f"{HIDDEN_TAG_PREFIX}pex_metadata"
+PEX_TAG_TAG = f"{SYSTEM_TAG_PREFIX}pex_tag"